+++ src/classes/Vector2.ts
@@ -6,6 +6,8 @@
 */
 export default class Vector2 {
     private coords: string;
+
+    public static readonly SIZE = 4;
     
     constructor(private readonly x: number, private readonly y: number) {
         this.coords = String.fromCodePoint(x) + String.fromCodePoint(y);