+++ src/errors/map/InvalidChunkDataError.ts
@@ -1,0 +1,5 @@
+export default class InvalidChunkDataError extends Error {
+    constructor() {
+        super("Invalid Chunk Data input.");
+    }
+}