+++ src/map/MapChunk.ts
@@ -2,7 +2,8 @@ import MapTile from "./MapTile";
 import Vector2 from "../classes/Vector2";
 
 export default class MapChunk {
-    private static readonly SIZE = 16;
+    public static readonly SIZE = 16;
+    public static readonly DATA_STRING_SIZE = MapChunk.SIZE * MapChunk.SIZE * 2 + Vector2.SIZE;
 
     constructor(private readonly tiles: MapTile[][], private readonly coords: Vector2) {
         this.tiles = tiles;